fix(bigquery): serialize datetime.time safely to avoid TypeError
CHATR-100/106/107/108

- Move CustomJSONEncoder to shared src/utils/json_utils.py
- save_response_in_bq: use CustomJSONEncoder in json.dumps(data, ...)
- get_bigquery_result: convert datetime.time (not just datetime/date) to
  ISO string, fixing the root source of the unserializable object
- error_interceptor: use CustomJSONEncoder when serializing input_body,
  so reporting an error no longer fails with the same TypeError
- Add regression tests covering all three fix points
